Added some tweaks

+++ backend/app/main.py
@@ -9,54 +9,94 @@ from dotenv import load_dotenv, dotenv_values
 import logging
 from pathlib import Path
 import anndata as ad
-from math import ceil
+from math import ceil, log10
 import dask.array
 import zarr
 import json
-
-# from serialize_util import convert_to_serializable
-
+import time
 
 load_dotenv()
-app = FastAPI()
+FRONTEND_ENDPOINT = os.environ.get("FRONTEND_ENDPOINT")
+ORIGINS = [
+    "https://tepohi.no",
+    "https://www.tepohi.no",
+    FRONTEND_ENDPOINT,
+]
 config = dotenv_values(".env")
+app = FastAPI()
+
+print(config, ORIGINS)
+
+app.add_middleware(
+    CORSMiddleware,
+    allow_origins=ORIGINS,
+    allow_credentials=True,
+    allow_methods=["*"],        # Allow all methods (GET, POST, etc.)
+    allow_headers=["*"],        # Allow all headers
+)
+
 adata = None
 
 UPLOAD_DIR = "/persistent01"
-FRONTEND_ENDPOINT = os.environ.get("FRONTEND_ENDPOINT")
 ADATA_CHUNK_SIZE = 1000
 UPLOAD_CHUNK_SIZE = 1024 * 1024 * 10 # 1MB chunks
 
 logger = logging.getLogger("uvicorn.error")
 logger.setLevel(logging.DEBUG)
 
-logger.debug(["config", FRONTEND_ENDPOINT])
+logger.debug(["config", ORIGINS])
 
 # Creates directory if it does not exist
 Path(UPLOAD_DIR).mkdir(parents=True, exist_ok=True)
 
-app.add_middleware(
-    CORSMiddleware,
-    allow_origins=[
-        # FRONTEND_ENDPOINT,
-        "*"
-    ],
-    allow_credentials=True,
-    allow_methods=["*"],        # Allow all methods (GET, POST, etc.)
-    allow_headers=["*"],        # Allow all headers
-)
 
 @app.get("/")
 async def root():
+  
+  result = []
+  file_sizes = [
+    "KB",
+    "MB",
+    "GB"
+  ]
+  
+  for file_id in os.listdir(UPLOAD_DIR):
+    
+    path = os.path.join(UPLOAD_DIR, file_id)
+    is_file = os.path.isfile(path)
+    
+    file_size_bytes = None
+    
+    if (is_file):
+      file_size_bytes = os.path.getsize(path)
+    else:
+      zarr_data = zarr.open_group(path)
+      store = zarr_data.store
+      file_size_bytes = sum(store.getsize(k) for k in store.keys())
+    
+    ceil_log_size = ceil(log10(file_size_bytes) / 4)
+    file_size = file_size_bytes * ((1/1024)**ceil_log_size)
+    file_size_result = f"{file_size:.2f} {file_sizes[ceil_log_size-1]}"
+    
+    result.append({
+      "name": file_id,
+      "file_size": file_size_result
+    })
+    
   return JSONResponse(content={
       "message": "Thesis API",
-      "files": list(filter(lambda x: x.endswith("zarr"), os.listdir(UPLOAD_DIR)))
+      "files": result
   })
 
 
 @app.get("/get_filenames")
 async def get_filenames():
-  return list(filter(lambda x: x.endswith("zarr"), os.listdir(UPLOAD_DIR)))
+  return JSONResponse(
+    content={
+      # "files": list(filter(lambda x: x.endswith("zarr"), os.listdir(UPLOAD_DIR)))
+      "files": list(filter(lambda x: x.endswith("zarr"), os.listdir(UPLOAD_DIR)))
+    }
+  )
 
 
 @app.get("/get_file_size")
@@ -183,7 +223,29 @@ async def get_gene_expression(file_id: str):
   
   
   
-
+@app.get("/measure_access_time")
+async def measure_access_time(file_id: str, attr: str):
+  
+  file_path = os.path.join(UPLOAD_DIR, file_id)
+  file_exists = os.path.exists(file_path)
+  
+  if (not file_exists):
+    return JSONResponse(content={
+      "file_path": file_path,
+      "file_exists": file_exists
+    })
+  
+  time_start = time.time()
+  zarr_data = zarr.open(file_path, "r")
+  time_end = time.time()
+  elapsed_time = time_end - time_start
+  
+  print(zarr_data)
+  
+  return JSONResponse(content={
+    "file_path": file_path,
+    "elapsed_time": elapsed_time
+  })
 
 @app.post("/upload_file_chunk/")
 async def upload_chunk(
@@ -226,7 +288,7 @@ async def upload_chunk(
   )
   
 @app.get("/assemble_file")
-def assemble_file(file_id: str):
+def assemble_file(file_id: str, total_chunks: int):
   
   dir_path = os.path.join(UPLOAD_DIR, file_id)
   
@@ -246,6 +308,16 @@ def assemble_file(file_id: str):
   dir_contents = sorted(os.listdir(dir_path), key=lambda x: int(x.split("_chunk_")[1]))
   
   dir_contents_len = len(dir_contents)
+  
+  if (dir_contents_len < total_chunks):
+    return JSONResponse(
+      content={
+        "message": {
+          "chunks_needed": total_chunks,
+          "chunks": dir_contents_len
+        }
+      }
+    )
   
   assembled_file_path = os.path.join(UPLOAD_DIR, f"{file_id}_assembled")
   
@@ -273,7 +345,7 @@ def assemble_file(file_id: str):
 
 
 @app.get("/convert_file")
-async def convert_h5ad_to_zarr(file_id: str):
+async def convert_h5ad_to_zarr(file_id: str, delete_after: bool=False):
   """
   Receives and assembles a chunk-separated file uploaded from the thesis frontend application :)
   """
@@ -315,7 +387,8 @@ async def convert_h5ad_to_zarr(file_id: str):
       os.path.join(UPLOAD_DIR, f"{file_name}-{file_id}.zarr")
     )
     
-    os.remove(file_path)
+    if (delete_after):
+      os.remove(file_path)
     
     return JSONResponse(content={"message": f"File converted to zarr"})
